Streamlined Code. Made plotting and metric-calculating functions
Created a function to calculate the mean and max signal of each ROI, reducing ever trial into a single number.

Created a function that plots these single values as a function of either the spatial period or the temporal frequency.

I also removed old helper methods and commented debugging lines